pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["whoopdata"]

[project]
name = "whoop-data"
version = "1.4.0"
description = "WHOOP and Withings Health Data Integration Platform"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Asif Laldin", email = "your.email@example.com" }
]
keywords = ["health", "data", "whoop", "withings", "api", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    # Core FastAPI dependencies
    "fastapi==0.115.12",
    "uvicorn==0.34.2",
    "python-multipart==0.0.20",
    "starlette==0.46.2",
    
    # Database
    "psycopg2-binary==2.9.10",
    "SQLAlchemy==2.0.40",
    
    # HTTP clients
    "httpx==0.28.1",
    "requests==2.32.3",
    
    # Environment and config
    "python-dotenv==1.0.1",
    "pydantic==2.10.5",
    "pydantic_core==2.27.2",
    
    # Data processing
    "numpy>=1.23.5,<2.0.0",
    "pandas==2.2.3",
    "python-dateutil==2.9.0.post0",
    
    # Visualization
    "matplotlib==3.9.3",
    "seaborn==0.13.2",
    
    # Scientific computing
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "xgboost>=2.0.0",
    
    # LangChain ecosystem
    "langchain==0.3.17",
    "langchain-openai==0.2.14",
    "langgraph==0.2.62",
    "langsmith==0.2.10",
    "langchain-experimental==0.3.4",
    "langchain-core>=0.3.33,<0.4.0",
    "langchain-text-splitters==0.3.4",
    "langchain-community==0.3.16",
    
    # OpenAI
    "openai==1.59.6",
    
    # Gradio for chat interface
    "gradio==5.9.1",
    
    # Utility
    "PyYAML==6.0.2",
    "Jinja2==3.1.5",
    
    # CLI utilities
    "rich",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
whoop-start = "whoopdata.cli:main"
whoop-etl = "whoopdata.etl:run_complete_etl"

[project.urls]
Homepage = "https://github.com/ald0405/whoop-data"
"Bug Reports" = "https://github.com/ald0405/whoop-data/issues"
Source = "https://github.com/ald0405/whoop-data"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.coverage.run]
source = ["whoopdata"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
